Move EnterPostBase to new file MarkdownInput
